First pass on a System Shock 1 object list for unity.

+++ gameobjectsrender.cpp
@@ -63,9 +63,9 @@ void CalcObjectXYZ(int game, float *offX, float *offY, float *offZ, tile LevelIn
 
 void CalcObjectXYZ(int game, float *offX, float *offY, float *offZ, tile LevelInfo[64][64], ObjectItem objList[1600], long nextObj, int x, int y, short WallAdjust)
 	{
-	int ResolutionXY = 7;	// A tile has a 7x7 grid for object positioning.
-	int ResolutionZ = 128;	//UW has 127 posible z positions for an object in tile.
-	if (game == SHOCK){ ResolutionXY = 256; ResolutionZ = 256; }	//Shock has more "z" in it.
+	float ResolutionXY = 7.0;	// A tile has a 7x7 grid for object positioning.
+	float ResolutionZ = 128.0;	//UW has 127 posible z positions for an object in tile.
+	if (game == SHOCK){ ResolutionXY = 256.0; ResolutionZ = 256.0; }	//Shock has more "z" in it.
 
 	*offX = 0;  *offY = 0; *offZ = 0;
